mtn v3 - refactor
