+++ maps/mountain_fortress_v3/mystical_chest.lua
@@ -1,9 +1,18 @@
+local Color = require 'utils.color_presets'
+local Event = require 'utils.event'
+local WPT = require 'maps.mountain_fortress_v3.table'
+local WD = require 'modules.wave_defense.table'
+local RPG = require 'modules.rpg.main'
+local Alert = require 'utils.alert'
+local Task = require 'utils.task'
+local Token = require 'utils.token'
 local Public = {}
 
 local shuffle = table.shuffle_table
-local insert = table.insert
 local random = math.random
 local floor = math.floor
+local insert = table.insert
+local abs = math.abs
 
 local item_worths = {
     ['wooden-chest'] = 4,
@@ -198,6 +207,340 @@ local function get_raffle_keys()
     return raffle_keys
 end
 
+local function get_random_weighted(weighted_table, weight_index)
+    local total_weight = 0
+    weight_index = weight_index or 1
+
+    for _, w in pairs(weighted_table) do
+        total_weight = total_weight + w[weight_index]
+    end
+
+    local index = random() * total_weight
+    local weight_sum = 0
+    for k, w in pairs(weighted_table) do
+        weight_sum = weight_sum + w[weight_index]
+        if weight_sum >= index then
+            return w, k
+        end
+    end
+end
+
+local function init_price_check(locomotive, mystical_chest)
+    local roll = 48 + abs(locomotive.position.y) * 1.75
+    roll = roll * random(25, 1337) * 0.01
+
+    local item_stacks = {}
+    local roll_count = 2
+    for _ = 1, roll_count, 1 do
+        for _, stack in pairs(Public.roll(floor(roll / roll_count), 2)) do
+            if not item_stacks[stack.name] then
+                item_stacks[stack.name] = stack.count
+            else
+                item_stacks[stack.name] = item_stacks[stack.name] + stack.count
+            end
+        end
+    end
+
+    local price = {}
+    for k, v in pairs(item_stacks) do
+        insert(price, {name = k, count = v})
+    end
+
+    mystical_chest.price = price
+end
+
+local function roll_item_stacks(remaining_budget, max_slots, blacklist)
+    local item_stack_set = {}
+    local item_stack_set_worth = 0
+
+    for i = 1, max_slots, 1 do
+        if remaining_budget <= 0 then
+            break
+        end
+        local item_stack = Public.roll_item_stack(remaining_budget, blacklist)
+        item_stack_set[i] = item_stack
+        remaining_budget = remaining_budget - item_stack.count * item_worths[item_stack.name]
+        item_stack_set_worth = item_stack_set_worth + item_stack.count * item_worths[item_stack.name]
+    end
+
+    return item_stack_set, item_stack_set_worth
+end
+
+local pause_wd_token =
+    Token.register(
+    function()
+        WD.pause(false)
+        local mc_rewards = WPT.get('mc_rewards')
+        mc_rewards.temp_boosts.wave_defense = nil
+    end
+)
+
+local restore_mining_speed_token =
+    Token.register(
+    function()
+        local mc_rewards = WPT.get('mc_rewards')
+        local force = game.forces.player
+        if mc_rewards.temp_boosts.mining then
+            force.manual_mining_speed_modifier = force.manual_mining_speed_modifier - 0.5
+            mc_rewards.temp_boosts.mining = nil
+        end
+    end
+)
+
+local restore_movement_speed_token =
+    Token.register(
+    function()
+        local mc_rewards = WPT.get('mc_rewards')
+        local force = game.forces.player
+        if mc_rewards.temp_boosts.movement then
+            force.character_running_speed_modifier = force.character_running_speed_modifier - 0.2
+            mc_rewards.temp_boosts.movement = nil
+        end
+    end
+)
+
+local mc_random_rewards = {
+    {
+        name = 'XP',
+        color = {r = 0.00, g = 0.45, b = 0.00},
+        tooltip = 'Selecting this will insert random XP onto the global xp pool!',
+        func = (function(player)
+            local rng = random(1024, 10240)
+            RPG.add_to_global_pool(rng)
+            local message = ({'locomotive.xp_bonus', player.name})
+            Alert.alert_all_players(15, message, nil, 'achievement/tech-maniac')
+            return true
+        end),
+        512
+    },
+    {
+        name = 'Coins',
+        color = {r = 0.00, g = 0.35, b = 0.00},
+        tooltip = 'Selecting this will grant each player some coins!',
+        func = (function(p)
+            local rng = random(256, 512)
+            local players = game.connected_players
+            for i = 1, #players do
+                local player = players[i]
+                if player and player.valid then
+                    if player.can_insert({name = 'coin', count = rng}) then
+                        player.insert({name = 'coin', count = rng})
+                    end
+                end
+            end
+            local message = ({'locomotive.coin_bonus', p.name})
+            Alert.alert_all_players(15, message, nil, 'achievement/tech-maniac')
+            return true
+        end),
+        512
+    },
+    {
+        name = 'Movement bonus',
+        str = 'movement',
+        color = {r = 0.00, g = 0.25, b = 0.00},
+        tooltip = 'Selecting this will grant the team a bonus movement speed for 15 minutes!',
+        func = (function(player)
+            local mc_rewards = WPT.get('mc_rewards')
+            local force = game.forces.player
+            if mc_rewards.temp_boosts.movement then
+                return false, '[Rewards] Movement bonus is already applied. Please choose another reward.'
+            end
+
+            mc_rewards.temp_boosts.movement = true
+
+            Task.set_timeout_in_ticks(54000, restore_movement_speed_token)
+            force.character_running_speed_modifier = force.character_running_speed_modifier + 0.2
+            local message = ({'locomotive.movement_bonus', player.name})
+            Alert.alert_all_players(15, message, nil, 'achievement/tech-maniac')
+            return true
+        end),
+        512
+    },
+    {
+        name = 'Mining bonus',
+        str = 'mining',
+        color = {r = 0.00, g = 0.00, b = 0.25},
+        tooltip = 'Selecting this will grant the team a bonus mining speed for 15 minutes!',
+        func = (function(player)
+            local mc_rewards = WPT.get('mc_rewards')
+            local force = game.forces.player
+            if mc_rewards.temp_boosts.mining then
+                return false, '[Rewards] Mining bonus is already applied. Please choose another reward.'
+            end
+
+            mc_rewards.temp_boosts.mining = true
+
+            Task.set_timeout_in_ticks(54000, restore_mining_speed_token)
+            force.manual_mining_speed_modifier = force.manual_mining_speed_modifier + 0.5
+            local message = ({'locomotive.mining_bonus', player.name})
+            Alert.alert_all_players(15, message, nil, 'achievement/tech-maniac')
+            return true
+        end),
+        512
+    },
+    {
+        name = 'Inventory Bonus',
+        color = {r = 0.00, g = 0.00, b = 0.25},
+        tooltip = 'Selecting this will grant the team permanent inventory bonus!',
+        func = (function(player)
+            local force = game.forces.player
+            force.character_inventory_slots_bonus = force.character_inventory_slots_bonus + 1
+            local message = ({'locomotive.inventory_bonus', player.name})
+            Alert.alert_all_players(15, message, nil, 'achievement/tech-maniac')
+            return true
+        end),
+        512
+    },
+    {
+        name = 'Heal Locomotive',
+        color = {r = 0.00, g = 0.00, b = 0.25},
+        tooltip = 'Selecting this will heal the main locomotive to full health!',
+        func = (function(player)
+            local locomotive_max_health = WPT.get('locomotive_max_health')
+            WPT.set('locomotive_health', locomotive_max_health)
+            local message = ({'locomotive.locomotive_health', player.name})
+            Alert.alert_all_players(15, message, nil, 'achievement/tech-maniac')
+            return true
+        end),
+        256
+    },
+    {
+        name = 'Wave Defense',
+        str = 'wave_defense',
+        color = {r = 0.35, g = 0.00, b = 0.00},
+        tooltip = 'Selecting this will pause the wave defense for 15 minutes. Ideal if you want to take a break!',
+        func = (function(player)
+            local mc_rewards = WPT.get('mc_rewards')
+            if mc_rewards.temp_boosts.wave_defense then
+                return false, '[Rewards] Wave Defense break is already applied. Please choose another reward.'
+            end
+            mc_rewards.temp_boosts.wave_defense = true
+
+            WD.pause(true)
+            Task.set_timeout_in_ticks(54000, pause_wd_token)
+            local message = ({'locomotive.wd_paused', player.name})
+            Alert.alert_all_players(15, message, nil, 'achievement/tech-maniac')
+            return true
+        end),
+        64
+    }
+}
+
+local function mystical_chest_reward(player)
+    if player.gui.screen['reward_system'] then
+        player.gui.screen['reward_system'].destroy()
+        return
+    end
+
+    local frame = player.gui.screen.add {type = 'frame', caption = 'Mystical Reward:', name = 'reward_system', direction = 'vertical'}
+    frame.auto_center = true
+    frame = frame.add {type = 'frame', name = 'reward_system_1', direction = 'vertical', style = 'inside_shallow_frame'}
+    frame.style.padding = 4
+
+    local mc_rewards = WPT.get('mc_rewards')
+    mc_rewards.current = {}
+
+    for i = 1, 3 do
+        local d, key = get_random_weighted(mc_random_rewards)
+        if not mc_rewards.current[key] and not mc_rewards.temp_boosts[d.str] then
+            mc_rewards.current[key] = {
+                id = i,
+                name = d.name
+            }
+            local b = frame.add({type = 'button', name = tostring(i), caption = d.name})
+            b.style.font_color = d.color
+            b.style.font = 'heading-2'
+            b.style.minimal_width = 180
+            b.style.horizontal_align = 'center'
+            b.tooltip = d.tooltip
+        end
+    end
+    if not next(mc_rewards.current) then
+        if player.gui.screen['reward_system'] then
+            player.gui.screen['reward_system'].destroy()
+        end
+        return player.print('[Rewards] No rewards are available.', Color.fail)
+    end
+
+    -- something fancy to reward players
+end
+
+local function container_opened(event)
+    local entity = event.entity
+    if not entity then
+        return
+    end
+    if not entity.valid then
+        return
+    end
+    if not entity.unit_number then
+        return
+    end
+
+    local mystical_chest = WPT.get('mystical_chest')
+    if not mystical_chest then
+        return
+    end
+    if not (mystical_chest.entity and mystical_chest.entity.valid) then
+        return
+    end
+
+    if entity.unit_number ~= mystical_chest.entity.unit_number then
+        return
+    end
+
+    local player = game.get_player(event.player_index)
+    if not (player and player.valid) then
+        return
+    end
+
+    Public.add_mystical_chest(player)
+end
+
+local function on_gui_opened(event)
+    container_opened(event)
+end
+
+local function on_gui_closed(event)
+    container_opened(event)
+end
+
+local function on_gui_click(event)
+    local element = event.element
+    if not (element and element.valid) then
+        return
+    end
+
+    if element.type ~= 'button' then
+        return
+    end
+
+    if element.parent.name ~= 'reward_system_1' then
+        return
+    end
+    local i = tonumber(element.name)
+
+    local mc_rewards = WPT.get('mc_rewards')
+    local current = mc_rewards.current
+
+    local player = game.get_player(element.player_index)
+    if not (player and player.valid) then
+        return
+    end
+
+    for id, data in pairs(current) do
+        if data.id == i then
+            local success, msg = mc_random_rewards[id].func(player)
+            if not success then
+                return player.print(msg, Color.fail)
+            end
+            break
+        end
+    end
+
+    element.parent.parent.destroy()
+end
+
 function Public.roll_item_stack(remaining_budget, blacklist)
     if remaining_budget <= 0 then
         return
@@ -227,23 +570,6 @@ function Public.roll_item_stack(remaining_budget, blacklist)
     end
 
     return {name = item_name, count = item_count}
-end
-
-local function roll_item_stacks(remaining_budget, max_slots, blacklist)
-    local item_stack_set = {}
-    local item_stack_set_worth = 0
-
-    for i = 1, max_slots, 1 do
-        if remaining_budget <= 0 then
-            break
-        end
-        local item_stack = Public.roll_item_stack(remaining_budget, blacklist)
-        item_stack_set[i] = item_stack
-        remaining_budget = remaining_budget - item_stack.count * item_worths[item_stack.name]
-        item_stack_set_worth = item_stack_set_worth + item_stack.count * item_worths[item_stack.name]
-    end
-
-    return item_stack_set, item_stack_set_worth
 end
 
 function Public.roll(budget, max_slots, blacklist)
@@ -278,5 +604,58 @@ function Public.roll(budget, max_slots, blacklist)
     end
     return final_stack_set
 end
+
+function Public.add_mystical_chest(player)
+    local locomotive = WPT.get('locomotive')
+    if not locomotive then
+        return
+    end
+    if not locomotive.valid then
+        return
+    end
+
+    local mystical_chest = WPT.get('mystical_chest')
+    if not (mystical_chest.entity and mystical_chest.entity.valid) then
+        return
+    end
+
+    if not mystical_chest.price then
+        init_price_check(locomotive, mystical_chest)
+    end
+
+    local entity = mystical_chest.entity
+
+    local inventory = mystical_chest.entity.get_inventory(defines.inventory.chest)
+
+    for key, item_stack in pairs(mystical_chest.price) do
+        local count_removed = inventory.remove(item_stack)
+        mystical_chest.price[key].count = mystical_chest.price[key].count - count_removed
+        if mystical_chest.price[key].count <= 0 then
+            table.remove(mystical_chest.price, key)
+        end
+    end
+
+    if #mystical_chest.price == 0 then
+        init_price_check(locomotive, mystical_chest)
+        if player and player.valid then
+            mystical_chest_reward(player)
+        end
+        return true
+    end
+
+    for slot = 1, 30, 1 do
+        entity.clear_request_slot(slot)
+    end
+
+    for slot, item_stack in pairs(mystical_chest.price) do
+        mystical_chest.entity.set_request_slot(item_stack, slot)
+    end
+end
+
+Public.init_price_check = init_price_check
+
+Event.add(defines.events.on_gui_opened, on_gui_opened)
+Event.add(defines.events.on_gui_closed, on_gui_closed)
+Event.add(defines.events.on_gui_click, on_gui_click)
 
 return Public
